+++ NeuralLayer.cpp
@@ -1,6 +1,6 @@
 #include "NeuralLayer.h"
 
-NeuralLayer::NeuralLayer(int input_size, int layer_size, int (*activation)(Neuron)) : input_size(input_size),
+NeuralLayer::NeuralLayer(int input_size, int layer_size, double (*activation)(Neuron)) : input_size(input_size),
                                                                                       layer_size(layer_size),
                                                                                       activation(activation) {
     neurons = new Neuron *[layer_size];
